+++ rsocket/handlers/request_response_requester.py
@@ -1,0 +1,28 @@
+from asyncio import Future
+
+from rsocket.exceptions import RSocketApplicationError
+from rsocket.frame import ErrorFrame, RequestResponseFrame, PayloadFrame, Frame
+from rsocket.handlers.stream_handler import StreamHandler
+from rsocket.payload import Payload
+
+
+class RequestResponseRequester(StreamHandler, Future):
+    def __init__(self, stream: int, socket, payload: Payload):
+        super().__init__(stream, socket)
+        request = RequestResponseFrame()
+        request.stream_id = self.stream
+        request.data = payload.data
+        request.metadata = payload.metadata
+        self.socket.send_frame(request)
+
+    async def frame_received(self, frame: Frame):
+        if isinstance(frame, PayloadFrame):
+            self.set_result(Payload(frame.data, frame.metadata))
+            self.socket.finish_stream(self.stream)
+        elif isinstance(frame, ErrorFrame):
+            self.set_exception(RSocketApplicationError(frame.data))
+            self.socket.finish_stream(self.stream)
+
+    def cancel(self, *args, **kwargs):
+        super().cancel(*args, **kwargs)
+        self.send_cancel()
