+++ rsocket/handlers/stream.py
@@ -1,0 +1,7 @@
+import abc
+
+
+class Stream(metaclass=abc.ABCMeta):
+    @abc.abstractmethod
+    def limit_rate(self, n: int):
+        ...
